Remove purgeCSS
Purgecss doesn't play nice with dynamic css classes and forces us to maintain a `whitelistPatterns`.

As of today CSS is still very light, arround 30kb gzipped. PurgeCSS brings down this number to 20kb which is a huge gain in relative terms but not very significant in absolute terms.

Whitelist would have been:
```
      whitelistPatterns: () => [/^icon-/, /^is-/, /^Select/, /^react-select/],
      whitelistPatternsChildren: () => [/content/, /video/]
```